dataloader/requirements.txt
requests
pillow
boto3
